example/requirements.txt
Django==1.5.1
Pillow==2.0.0
argparse==1.2.1
django-autoslug==1.6.1
django-debug-toolbar==0.9.4
django-lime==0.2
django-reversion==1.6.5
django-tinymce==1.5.1
django-werewolf
ipdb==0.7
ipython==0.13.2
python-dateutil==2.1
pytz==2013b
six==1.3.0
wsgiref==0.1.2
